refactor requestOptions
